Fixed display placing and added full control over buttons and entry field
Display centering that was previously forgotten to be enabled now is.
Instead of using various parameters to control button design it now uses
a dictionary that gets directly passed to the buttons and the entryfield
to give full control over them.
Keysize now will be auto-calculated when either it is not set or would
produce buttons too big to display.